+++ CHANGES.md
@@ -3,6 +3,8 @@ Changes
 
 0.3.2 (unreleased)
 
+- Add check ScopeCount
+
 
 0.3.1 (2019-09-22)